refactor: split puzzle-generation provider and parsing
